Move MailBuilder related tests to own testcase
