+++ tests/Mail/MailMessageTest.php
@@ -14,7 +14,6 @@
 namespace Eventum\Test\Mail;
 
 use Date_Helper;
-use Eventum\Mail\MailBuilder;
 use Eventum\Mail\MailMessage;
 use Eventum\Test\TestCase;
 use Laminas\Mail\AddressList;
@@ -612,34 +611,6 @@ class MailMessageTest extends TestCase
         $mail->addHeaders($headers);
         $result = $mail->messageId;
         $this->assertEquals($messageId, $result);
-    }
-
-    public function testZFPlainMail(): void
-    {
-        $text_message = 'zzzxx';
-        $issue_id = 1;
-        $from = '"Admin User " <note-3@eventum.example.org>';
-        $to = '"Admin User" <admin@example.com>';
-        $subject = '[#3] Note: Re: pläh';
-        $type = 'assignment';
-
-        // use mail builder
-        $builder = new MailBuilder();
-        $builder
-            ->addTextPart($text_message)
-            ->getMessage()
-            ->setFrom($from)
-            ->setTo($to)
-            ->setSubject($subject);
-        $mail = $builder->toMailMessage();
-
-        $options = [
-            'save_email_copy' => true,
-            'issue_id' => $issue_id,
-            'type' => $type,
-        ];
-        Mail_Queue::queue($mail, $to, $options);
-        $this->assertTrue(true);
     }
 
     /**
